requirements.txt
absl-py==0.11.0
astunparse==1.6.3
cachetools==4.1.1
certifi==2020.6.20
cloudpickle==1.6.0
cycler==0.10.0
gast==0.3.3
google-auth==1.23.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.33.2
gym==0.17.3
h5py==2.10.0
joblib==0.16.0
Keras==2.4.3
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
Markdown==3.3.3
matplotlib==3.3.2
mkl-fft==1.2.0
mkl-random==1.1.1
mkl-service==2.3.0
numpy==1.18.5
olefile==0.46
opencv-python==4.4.0.46
opt-einsum==3.3.0
pandas==1.1.2
Pillow @ file:///tmp/build/80754af9/pillow_1603822255246/work
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydicom==2.1.1
pydot==1.4.1
pyglet==1.5.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
requests-oauthlib==1.3.0
rsa==4.6
scikit-learn==0.23.2
scikit-plot==0.3.7
scipy==1.5.2
seaborn==0.11.0
six @ file:///tmp/build/80754af9/six_1605205327372/work
sklearn==0.0
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.1
tensorflow-estimator==2.3.0
termcolor==1.1.0
threadpoolctl==2.1.0
Werkzeug==1.0.1
wrapt==1.12.1
xlrd==1.2.0
